requirements.txt
# Python depenencies

dnspython==2.7.0
logging==0.4.9.6
pymongo==4.10.1
python-dotenv==1.0.1
